+++ CHANGELOG.md
@@ -8,6 +8,7 @@
     * `block::SetAttribute` is a modifier which overrides a single attribute of a block, instead of all of them.
       It replaces `Modifier::Attributes`.
     * `math::Vol::origin_empty()`, infallible construction of empty volume.
+    * `time::Clock::phase()`, returns the clock’s phase, which was previously only obtainable through `Tick`.
     * `time::Tick::{delta_t_f64, delta_t_ps64}()`, convenience for common conversions.
 
 * `all-is-cubes-render` library: